Refactor: 호스트 주소등록 리팩토링
