+++ src/hooks/useFetch.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function useFetch() {
+  return <div>useFetch</div>;
+}
+
+export default useFetch;
