+++ src/pages/HostRegister/HostRegisterStep/DetailInfo/useDetailInfo.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function useDetailInfo() {
+  return <div>useDetailInfo</div>;
+}
+
+export default useDetailInfo;
